requirements.txt
requests~=2.31.0
pandas
geopandas~=0.14.3
SQLAlchemy~=2.0.28
psycopg2-binary

pymongo~=4.6.2
pyarrow~=15.0.1
sqlalchemy-timescaledb

python-snappy
pyspark
delta-spark
motion-lake-client